Split runner shell job execution tests
